+++ jarchive.el
@@ -163,22 +163,43 @@ handle it. If it is not a jar call ORIGINAL-FN."
 
 ;;;###autoload
 (defun jarchive-patch-eglot ()
-  "Patch old versions of Eglot to work with Jarchive."
+  "This should no longer be used.
+Recent versions of Eglot take care of this."
   (interactive) ;; TODO, remove when eglot is updated in melpa
   (unless (or (and (advice-member-p #'jarchive--wrap-legacy-eglot--path-to-uri 'eglot--path-to-uri)
                    (advice-member-p #'jarchive--wrap-legacy-eglot--uri-to-path 'eglot--uri-to-path))
               (<= 29 emacs-major-version))
     (advice-add 'eglot--path-to-uri :around #'jarchive--wrap-legacy-eglot--path-to-uri)
     (advice-add 'eglot--uri-to-path :around #'jarchive--wrap-legacy-eglot--uri-to-path)
-    (message "[jarchive] Eglot successfully patched.")))
+    (message "[jarchive] Eglot successfully patched."))
+  (warn "`jarchive-patch-eglot' is obsolete and should not be needed with
+recent versions of Eglot (including the version shipped in Emacs 29+)"))
+
+(make-obsolete 'jarchive-patch-eglot nil "0.11.0")
+
+(defvar jarchive--file-name-handler
+  (cons jarchive--uri-regex #'jarchive--file-name-handler))
 
 ;;;###autoload
-(defun jarchive-setup ()
-  "Setup jarchive, enabling Emacs to open files inside jar archives.
-the files can be identified with the `jar' uri scheme."
-  (interactive)
-  (add-to-list 'file-name-handler-alist (cons jarchive--uri-regex #'jarchive--file-name-handler))
-  (add-to-list 'find-file-not-found-functions #'jarchive--find-file-not-found))
+(define-minor-mode jarchive-mode
+  "Teach Emacs to work with jar URIs.
+Opening a jar URI via Emacs functions like `find-file' will
+automatically extract the contents of a file contained in the jar
+and open them in a buffer."
+  :global t
+  :lighter " Jar"
+  (if jarchive-mode
+      (progn
+        (add-to-list 'file-name-handler-alist jarchive--file-name-handler)
+        (add-to-list 'find-file-not-found-functions #'jarchive--find-file-not-found))
+    (setq file-name-handler-alist
+          (remove jarchive--file-name-handler file-name-handler-alist)
+          find-file-not-found-functions
+          (remove #'jarchive--find-file-not-found find-file-not-found-functions))))
+
+;;;###autoload
+(define-obsolete-function-alias
+  'jarchive-setup 'jarchive-mode "0.11.0" "Obsolete. Set up Jarchive to open jar URIs.")
 
 (provide 'jarchive)
 ;;; jarchive.el ends here